refactor: removed dead wood
