+++ src/frontend/fields/animation-field.story.vue
@@ -47,8 +47,8 @@
 
 <script setup lang="ts">
 import AnimationField from './animation-field.vue';
-import ErrorControl from './helpers/error-control.vue';
-import ModelControl from './helpers/model-control.vue';
+import ErrorControl from '../test/error-control.vue';
+import ModelControl from '../test/model-control.vue';
 import { useModelStore, useSharedStore } from '../store';
 import type { StoryHandler } from '../shared/helpers';
