+++ src/frontend/fields/image-field.story.vue
@@ -47,8 +47,8 @@
 
 <script setup lang="ts">
 import ImageField from './image-field.vue';
-import ErrorControl from './helpers/error-control.vue';
-import ModelControl from './helpers/model-control.vue';
+import ErrorControl from '../test/error-control.vue';
+import ModelControl from '../test/model-control.vue';
 import {
   objectErrors,
   objectModel,
