+++ src/frontend/fields/object-field.story.vue
@@ -55,8 +55,8 @@
 <script setup lang="ts">
 import { useModelStore, useSharedStore } from '../store';
 import ObjectField from './object-field.vue';
-import ErrorControl from './helpers/error-control.vue';
-import ModelControl from './helpers/model-control.vue';
+import ErrorControl from '../test/error-control.vue';
+import ModelControl from '../test/model-control.vue';
 import {
   objectErrors,
   objectModel,
