+++ src/frontend/fields/panel-field.story.vue
@@ -30,8 +30,8 @@
 
 <script lang="ts" setup>
 import PanelField from './panel-field.vue';
-import ModelControl from './helpers/model-control.vue';
-import ErrorControl from './helpers/error-control.vue';
+import ModelControl from '../test/model-control.vue';
+import ErrorControl from '../test/error-control.vue';
 import { useModelStore, useSharedStore } from '../store';
 
 import type { StoryHandler } from '../shared/helpers';
