+++ src/frontend/fields/scripture-field.story.vue
@@ -35,8 +35,8 @@
 
 <script setup lang="ts">
 import ScriptureField from './scripture-field.vue';
-import ErrorControl from './helpers/error-control.vue';
-import ModelControl from './helpers/model-control.vue';
+import ErrorControl from '../test/error-control.vue';
+import ModelControl from '../test/model-control.vue';
 import { scriptureSpec, scriptureModel } from '../test/mocks';
 
 import { useModelStore, useSharedStore } from '../store';
