+++ src/frontend/fields/video-field.story.vue
@@ -31,8 +31,8 @@
 
 <script setup lang="ts">
 import VideoField from './video-field.vue';
-import ErrorControl from './helpers/error-control.vue';
-import ModelControl from './helpers/model-control.vue';
+import ErrorControl from '../test/error-control.vue';
+import ModelControl from '../test/model-control.vue';
 import { useModelStore, useSharedStore } from '../store';
 import type { StoryHandler } from '../shared/helpers';
